+++ src/email-input/README.md
@@ -1,0 +1,3 @@
+# @dojo/widgets/email-input
+
+Dojo's `EmailInput` widget provides a basic email input widget.
